Added backup authentication method
If authentication is deactivated touchID would be not available in iOS 10. This plugin now used LAPolicyDeviceOwnerAuthentication which checks both TouchID and the device passcode. If the user has used to many attempts to unlock the app the device will be locked for a bit. After that specific time the TouchID is still locked. We use the devices passcode to unlock TouchID and ensure a secure authentication.